refactor fs code, move OS specific edge cases to fs-normalized.js
